Move MockKube from cluster-operator → test module
